Fair team distribution
Added a greedy sampling method that will try to balance the teams based on a predefined weight
for each value urn. For the 3 urns, the values 3, 2, 1 will be used where 3 is for the first urn, which
is the most valuable.

In the same commit, there was added a change where smaller teams will be filled with an 'X' as a
placeholder